% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_erhmm.R
\name{AERHMMClass}
\alias{AERHMMClass}
\title{ErlangHMM for MAP with fixed phases}
\description{
ErlangHMM for MAP with fixed phases

ErlangHMM for MAP with fixed phases
}
\details{
A special case of MAP.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AERHMMClass-alpha}{\code{AERHMMClass$alpha()}}
\item \href{#method-AERHMMClass-shape}{\code{AERHMMClass$shape()}}
\item \href{#method-AERHMMClass-rate}{\code{AERHMMClass$rate()}}
\item \href{#method-AERHMMClass-P}{\code{AERHMMClass$P()}}
\item \href{#method-AERHMMClass-xi}{\code{AERHMMClass$xi()}}
\item \href{#method-AERHMMClass-new}{\code{AERHMMClass$new()}}
\item \href{#method-AERHMMClass-copy}{\code{AERHMMClass$copy()}}
\item \href{#method-AERHMMClass-size}{\code{AERHMMClass$size()}}
\item \href{#method-AERHMMClass-df}{\code{AERHMMClass$df()}}
\item \href{#method-AERHMMClass-print}{\code{AERHMMClass$print()}}
\item \href{#method-AERHMMClass-mmoment}{\code{AERHMMClass$mmoment()}}
\item \href{#method-AERHMMClass-jmoment}{\code{AERHMMClass$jmoment()}}
\item \href{#method-AERHMMClass-acf}{\code{AERHMMClass$acf()}}
\item \href{#method-AERHMMClass-emfit}{\code{AERHMMClass$emfit()}}
\item \href{#method-AERHMMClass-init}{\code{AERHMMClass$init()}}
\item \href{#method-AERHMMClass-clone}{\code{AERHMMClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-alpha"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-alpha}{}}}
\subsection{Method \code{alpha()}}{
Get alpha
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$alpha()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of alpha
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-shape"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-shape}{}}}
\subsection{Method \code{shape()}}{
Get shape
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$shape()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of shapes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-rate"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-rate}{}}}
\subsection{Method \code{rate()}}{
Get rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$rate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of rates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-P"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-P}{}}}
\subsection{Method \code{P()}}{
Get P
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$P()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A matrix of P
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-xi"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-xi}{}}}
\subsection{Method \code{xi()}}{
Get exit rates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$xi()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of exit rates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-new}{}}}
\subsection{Method \code{new()}}{
Create an AERHMM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$new(size, erhmm)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{An integer of the number of phases}

\item{\code{erhmm}}{An instance of ERHMM}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of AERHMM
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-copy"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-copy}{}}}
\subsection{Method \code{copy()}}{
copy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$copy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-size"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-size}{}}}
\subsection{Method \code{size()}}{
The number of components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of components
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-df"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-df}{}}}
\subsection{Method \code{df()}}{
Degrees of freedom
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The degrees of freedom
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-print"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-print}{}}}
\subsection{Method \code{print()}}{
Print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-mmoment"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-mmoment}{}}}
\subsection{Method \code{mmoment()}}{
Marginal moments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$mmoment(k, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{An integer of degree}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of moments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-jmoment"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-jmoment}{}}}
\subsection{Method \code{jmoment()}}{
Joint moments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$jmoment(lag, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lag}}{An integer of lag}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix of moments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-acf"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-acf}{}}}
\subsection{Method \code{acf()}}{
k-lag correlation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$acf(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector for k-lag correlation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-emfit"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-emfit}{}}}
\subsection{Method \code{emfit()}}{
Run EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$emfit(data, options, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{options}}{A list of options}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-init"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-init}{}}}
\subsection{Method \code{init()}}{
Initialize with data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$init(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{...}}{Others}

\item{\code{options}}{A list of options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AERHMMClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AERHMMClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AERHMMClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
