% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_herlang.R
\name{AHerlang}
\alias{AHerlang}
\title{Hyper-Erlang distribution with a fixed phase}
\description{
Hyper-Erlang distribution with a fixed phase

Hyper-Erlang distribution with a fixed phase
}
\details{
A mixture of Erlang distributions. A subclass of PH distributions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AHerlang-alpha}{\code{AHerlang$alpha()}}
\item \href{#method-AHerlang-shape}{\code{AHerlang$shape()}}
\item \href{#method-AHerlang-rate}{\code{AHerlang$rate()}}
\item \href{#method-AHerlang-new}{\code{AHerlang$new()}}
\item \href{#method-AHerlang-size}{\code{AHerlang$size()}}
\item \href{#method-AHerlang-df}{\code{AHerlang$df()}}
\item \href{#method-AHerlang-moment}{\code{AHerlang$moment()}}
\item \href{#method-AHerlang-print}{\code{AHerlang$print()}}
\item \href{#method-AHerlang-pdf}{\code{AHerlang$pdf()}}
\item \href{#method-AHerlang-cdf}{\code{AHerlang$cdf()}}
\item \href{#method-AHerlang-ccdf}{\code{AHerlang$ccdf()}}
\item \href{#method-AHerlang-sample}{\code{AHerlang$sample()}}
\item \href{#method-AHerlang-emfit}{\code{AHerlang$emfit()}}
\item \href{#method-AHerlang-init}{\code{AHerlang$init()}}
\item \href{#method-AHerlang-clone}{\code{AHerlang$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-alpha"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-alpha}{}}}
\subsection{Method \code{alpha()}}{
Get alpha
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$alpha()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of alpha
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-shape"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-shape}{}}}
\subsection{Method \code{shape()}}{
Get shape
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$shape()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of shapes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-rate"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-rate}{}}}
\subsection{Method \code{rate()}}{
Get rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$rate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of rates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-new"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-new}{}}}
\subsection{Method \code{new()}}{
Create a hyper-Erlang distribution with fixed phases
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$new(size, herlang)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{An integer of the number of phases}

\item{\code{herlang}}{An instance of HErlang}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of AHerlang
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-size"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-size}{}}}
\subsection{Method \code{size()}}{
The number of components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of components
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-df"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-df}{}}}
\subsection{Method \code{df()}}{
Degrees of freedom
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The degrees of freedom
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-moment"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-moment}{}}}
\subsection{Method \code{moment()}}{
Moments of HErlang
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$moment(k, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{A value to indicate the degrees of moments. k-th moment}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of moments from 1st to k-th moments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-print"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-print}{}}}
\subsection{Method \code{print()}}{
Print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-pdf}{}}}
\subsection{Method \code{pdf()}}{
PDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$pdf(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of densities.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-cdf}{}}}
\subsection{Method \code{cdf()}}{
CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$cdf(q, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of probabilities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-ccdf"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-ccdf}{}}}
\subsection{Method \code{ccdf()}}{
Complementary CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$ccdf(q, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of probabilies
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-sample"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-sample}{}}}
\subsection{Method \code{sample()}}{
Make a sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$sample(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A sample of HErlang
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-emfit"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-emfit}{}}}
\subsection{Method \code{emfit()}}{
Run EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$emfit(data, options, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{options}}{A list of options}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-init"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-init}{}}}
\subsection{Method \code{init()}}{
Initialize with data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$init(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{...}}{Others}

\item{\code{options}}{A list of options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlang-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlang-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlang$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
