% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_herlang.R
\name{AHerlangClass}
\alias{AHerlangClass}
\title{Hyper-Erlang distribution with a fixed phase}
\description{
Hyper-Erlang distribution with a fixed phase

Hyper-Erlang distribution with a fixed phase
}
\details{
A mixture of Erlang distributions. A subclass of PH distributions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AHerlangClass-mixrate}{\code{AHerlangClass$mixrate()}}
\item \href{#method-AHerlangClass-shape}{\code{AHerlangClass$shape()}}
\item \href{#method-AHerlangClass-rate}{\code{AHerlangClass$rate()}}
\item \href{#method-AHerlangClass-new}{\code{AHerlangClass$new()}}
\item \href{#method-AHerlangClass-copy}{\code{AHerlangClass$copy()}}
\item \href{#method-AHerlangClass-size}{\code{AHerlangClass$size()}}
\item \href{#method-AHerlangClass-df}{\code{AHerlangClass$df()}}
\item \href{#method-AHerlangClass-moment}{\code{AHerlangClass$moment()}}
\item \href{#method-AHerlangClass-print}{\code{AHerlangClass$print()}}
\item \href{#method-AHerlangClass-pdf}{\code{AHerlangClass$pdf()}}
\item \href{#method-AHerlangClass-cdf}{\code{AHerlangClass$cdf()}}
\item \href{#method-AHerlangClass-ccdf}{\code{AHerlangClass$ccdf()}}
\item \href{#method-AHerlangClass-sample}{\code{AHerlangClass$sample()}}
\item \href{#method-AHerlangClass-emfit}{\code{AHerlangClass$emfit()}}
\item \href{#method-AHerlangClass-init}{\code{AHerlangClass$init()}}
\item \href{#method-AHerlangClass-clone}{\code{AHerlangClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-mixrate"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-mixrate}{}}}
\subsection{Method \code{mixrate()}}{
Get mixrate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$mixrate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of mixrate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-shape"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-shape}{}}}
\subsection{Method \code{shape()}}{
Get shape
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$shape()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of shapes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-rate"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-rate}{}}}
\subsection{Method \code{rate()}}{
Get rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$rate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of rates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-new}{}}}
\subsection{Method \code{new()}}{
Create a hyper-Erlang distribution with fixed phases
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$new(size, herlang)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{An integer of the number of phases}

\item{\code{herlang}}{An instance of HErlang}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of AHerlang
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-copy"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-copy}{}}}
\subsection{Method \code{copy()}}{
copy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$copy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-size"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-size}{}}}
\subsection{Method \code{size()}}{
The number of components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of components
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-df"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-df}{}}}
\subsection{Method \code{df()}}{
Degrees of freedom
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The degrees of freedom
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-moment"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-moment}{}}}
\subsection{Method \code{moment()}}{
Moments of HErlang
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$moment(k, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{A value to indicate the degrees of moments. k-th moment}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of moments from 1st to k-th moments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-print"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-print}{}}}
\subsection{Method \code{print()}}{
Print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-pdf}{}}}
\subsection{Method \code{pdf()}}{
PDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$pdf(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of densities.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-cdf}{}}}
\subsection{Method \code{cdf()}}{
CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$cdf(q, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of probabilities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-ccdf"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-ccdf}{}}}
\subsection{Method \code{ccdf()}}{
Complementary CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$ccdf(q, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of probabilities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-sample"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-sample}{}}}
\subsection{Method \code{sample()}}{
Make a sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$sample(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A sample of HErlang
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-emfit"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-emfit}{}}}
\subsection{Method \code{emfit()}}{
Run EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$emfit(data, options, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{options}}{A list of options}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-init"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-init}{}}}
\subsection{Method \code{init()}}{
Initialize with data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$init(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{...}}{Others}

\item{\code{options}}{A list of options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AHerlangClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AHerlangClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AHerlangClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
