% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_herlang.R
\name{as.gph}
\alias{as.gph}
\title{Convert from HErlang to GPH}
\usage{
as.gph(h)
}
\arguments{
\item{h}{An instance of HErlang}
}
\value{
An instance of GPH
}
\description{
Convert from hyper-Erlang distribution to the general PH distribution
}
\examples{
#' ## create a hyper Erlang with specific parameters
(param <- herlang(shape=c(2,3), mixrate=c(0.3,0.7), rate=c(1.0,10.0)))

## convert to a general PH
as.gph(param)

}
