% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_erhmm.R
\name{as.map}
\alias{as.map}
\title{Convert from ERHMM to MAP}
\usage{
as.map(x)
}
\arguments{
\item{x}{An instance of ERHMM}
}
\value{
An instance of MAP
}
\description{
Convert from ERHMM to the general MAP
}
\examples{
## create a hyper Erlang with specific parameters
(param <- erhmm(shape=c(2,3), alpha=c(0.3,0.7), rate=c(1.0,10.0)))

## convert to a general PH
as.map(param)

}
