% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cf1.R
\name{cf1.param.power}
\alias{cf1.param.power}
\title{Determine CF1 parameters}
\usage{
cf1.param.power(size, mean, s)
}
\arguments{
\item{size}{An integer of the number of phases}

\item{mean}{A value of mean of data}

\item{s}{A value of fraction of minimum and maximum rates}
}
\value{
A list of alpha and rate
}
\description{
Determine CF1 parameters based on the power rule.
}
