% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cf1.R
\name{CF1}
\alias{CF1}
\title{Canonical phase-type distribution}
\description{
Canonical phase-type distribution

Canonical phase-type distribution
}
\details{
A continuous distribution dominated by a continuous-time Markov chain.
A random time is given by an absorbing time. In the CF1 (canonical form 1),
the infinitesimal generator is given by a bi-diagonal matrix, and whose order
is determiend by the ascending order.
}
\section{Super class}{
\code{\link[mapfit:GPH]{mapfit::GPH}} -> \code{CF1}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rate}}{A vector of rates}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CF1-new}{\code{CF1$new()}}
\item \href{#method-CF1-print}{\code{CF1$print()}}
\item \href{#method-CF1-clone}{\code{CF1$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPH" data-id="ccdf"><a href='../../mapfit/html/GPH.html#method-GPH-ccdf'><code>mapfit::GPH$ccdf()</code></a></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPH" data-id="cdf"><a href='../../mapfit/html/GPH.html#method-GPH-cdf'><code>mapfit::GPH$cdf()</code></a></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPH" data-id="df"><a href='../../mapfit/html/GPH.html#method-GPH-df'><code>mapfit::GPH$df()</code></a></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPH" data-id="moment"><a href='../../mapfit/html/GPH.html#method-GPH-moment'><code>mapfit::GPH$moment()</code></a></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPH" data-id="pdf"><a href='../../mapfit/html/GPH.html#method-GPH-pdf'><code>mapfit::GPH$pdf()</code></a></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPH" data-id="sample"><a href='../../mapfit/html/GPH.html#method-GPH-sample'><code>mapfit::GPH$sample()</code></a></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPH" data-id="size"><a href='../../mapfit/html/GPH.html#method-GPH-size'><code>mapfit::GPH$size()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF1-new"></a>}}
\if{latex}{\out{\hypertarget{method-CF1-new}{}}}
\subsection{Method \code{new()}}{
Create a CF1
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF1$new(alpha, rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{A vector of initial probability}

\item{\code{rate}}{A vector of rates}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of CF1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF1-print"></a>}}
\if{latex}{\out{\hypertarget{method-CF1-print}{}}}
\subsection{Method \code{print()}}{
Print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF1$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CF1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
