% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{ctmc.st}
\alias{ctmc.st}
\title{Markov stationary}
\usage{
ctmc.st(Q)
}
\arguments{
\item{Q}{DTMC/CTMC kernel}
}
\value{
The stationary vector of DTMC/CTMC
}
\description{
Compute the stationary vector with GTH
}
