% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map.R
\name{data.frame.map.group}
\alias{data.frame.map.group}
\title{Create group data for map}
\usage{
data.frame.map.group(counts, breaks, intervals, instants)
}
\arguments{
\item{counts}{A vector of the number of samples}

\item{breaks}{A vector of break points}

\item{intervals}{A vector of differences of time}

\item{instants}{A vector meaning whether a sample is observed at the end of break.}
}
\value{
A dataframe
}
\description{
Provide the data.frame for group data.
}
\examples{
t <- c(1,1,1,1,1)
n <- c(1,3,0,0,1)

dat <- data.frame.map.group(counts=n, intervals=t)
mean(dat)
print(dat)

}
