% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map.R
\name{data.frame.map.time}
\alias{data.frame.map.time}
\title{Create data for map}
\usage{
data.frame.map.time(time, intervals)
}
\arguments{
\item{time}{A vector for cumulative time}

\item{intervals}{A vector for time intervals}
}
\value{
A dataframe
}
\description{
Provide a data.frame with samples.
}
\note{
\itemize{
\item If both time and intervals are used, time is used.
\item map.time is given by a special case of map.group.
}
}
\examples{
x <- runif(10)

dat <- data.frame.map.time(time=x)
mean(dat)
print(dat)

}
