% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_phase.R
\name{data.frame.phase.interval}
\alias{data.frame.phase.interval}
\title{Create interval data for phase}
\usage{
data.frame.phase.interval(data, weights)
}
\arguments{
\item{data}{A list of point and interval data}

\item{weights}{A vector of left-truncation time points. If tau is missing, all the left-truncation times are NA (no truncation).}
}
\value{
A dataframe
}
\description{
Provide the data.frame for interval data.
}
\examples{
dat <- data.frame.phase.interval(data=list(1.0, c(0.0, 1.0), c(1.0, Inf)), weights=c(1, 2.0, 1.0))
print(dat)
mean(dat)

}
