% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_phase.R
\name{data.frame.phase.surv}
\alias{data.frame.phase.surv}
\title{Create left-truncated and right-censored data for phase}
\usage{
data.frame.phase.surv(x, delta, tau)
}
\arguments{
\item{x}{A vector of time points}

\item{delta}{A vector of indicators whether x is censoring time or not. If delta=0, the corresponding x is the censoring time
If delta=1, the corresponding x is the event time.}

\item{tau}{A vector of left-truncation time points. If tau is missing, all the left-truncation times are NA (no truncation).}
}
\value{
A dataframe
}
\description{
Provide the data.frame for left-truncated and right-censored data.
}
\examples{
dat <- data.frame.phase.surv(x=c(1.0, 4.5, 1.2), delta=c(1, 0, 0), tau=c(NA, 2.0, NA))
print(dat)
mean(dat)

}
