% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_phase.R
\name{data.frame.phase.time}
\alias{data.frame.phase.time}
\title{Create data for phase with weighted sample}
\usage{
data.frame.phase.time(x, weights)
}
\arguments{
\item{x}{A vector of point (quantiles)}

\item{weights}{A vector of weights}
}
\value{
A dataframe
}
\description{
Provide a data.frame with weighted samples.
}
\note{
The point time is sorted and their differences are stored as the column of \code{time}
}
\examples{
x <- runif(10)
w <- runif(10)

dat <- data.frame.phase.time(x=x, weights=w)
print(dat)
mean(dat)

}
