% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{diag.padding.zero}
\alias{diag.padding.zero}
\title{Padding zero to omitted diagonal elements}
\usage{
diag.padding.zero(A)
}
\arguments{
\item{A}{A sparse matrix}
}
\value{
A sparse matrix whose all diagonal elements are not omitted even if
they have zero values.
}
\description{
This function provides a sparse matrix whose all diagonal elements are not
omitted even if they have zero values.
}
