% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ph.R
\name{dphase}
\alias{dphase}
\title{Probability density function of PH distribution}
\usage{
dphase(x, ph, log = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{ph}{An instance of PH distribution.}

\item{log}{logical; if TRUE, densities y are returned as log(y)}

\item{...}{Others.}
}
\value{
A vector of densities.
}
\description{
Compute the probability density function (p.d.f.) for a given PH distribution
}
\examples{
## create PH instance (ex. full PH with 5 phases)
(phdist <- ph(5))

x <- runif(10) # points of quantiles
dphase(x, ph)

}
