% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_erhmm.R
\name{erhmm.param}
\alias{erhmm.param}
\title{Determine ERHMM parameters}
\usage{
erhmm.param(data, skel, ...)
}
\arguments{
\item{data}{A dataframe}

\item{skel}{An instance of ERHMM used as a skeleton}

\item{...}{Others}
}
\value{
An instance of ERHMM
}
\description{
Determine ERHMM parameters with k-means.
}
