% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_erhmm.R
\name{ERHMMClass}
\alias{ERHMMClass}
\title{ErlangHMM for MAP}
\description{
ErlangHMM for MAP

ErlangHMM for MAP
}
\details{
A special case of MAP.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ERHMMClass-alpha}{\code{ERHMMClass$alpha()}}
\item \href{#method-ERHMMClass-shape}{\code{ERHMMClass$shape()}}
\item \href{#method-ERHMMClass-rate}{\code{ERHMMClass$rate()}}
\item \href{#method-ERHMMClass-P}{\code{ERHMMClass$P()}}
\item \href{#method-ERHMMClass-xi}{\code{ERHMMClass$xi()}}
\item \href{#method-ERHMMClass-new}{\code{ERHMMClass$new()}}
\item \href{#method-ERHMMClass-copy}{\code{ERHMMClass$copy()}}
\item \href{#method-ERHMMClass-size}{\code{ERHMMClass$size()}}
\item \href{#method-ERHMMClass-df}{\code{ERHMMClass$df()}}
\item \href{#method-ERHMMClass-print}{\code{ERHMMClass$print()}}
\item \href{#method-ERHMMClass-mmoment}{\code{ERHMMClass$mmoment()}}
\item \href{#method-ERHMMClass-jmoment}{\code{ERHMMClass$jmoment()}}
\item \href{#method-ERHMMClass-acf}{\code{ERHMMClass$acf()}}
\item \href{#method-ERHMMClass-emfit}{\code{ERHMMClass$emfit()}}
\item \href{#method-ERHMMClass-init}{\code{ERHMMClass$init()}}
\item \href{#method-ERHMMClass-clone}{\code{ERHMMClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-alpha"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-alpha}{}}}
\subsection{Method \code{alpha()}}{
Get alpha
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$alpha()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of alpha
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-shape"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-shape}{}}}
\subsection{Method \code{shape()}}{
Get shape
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$shape()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of shapes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-rate"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-rate}{}}}
\subsection{Method \code{rate()}}{
Get rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$rate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of rates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-P"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-P}{}}}
\subsection{Method \code{P()}}{
Get P
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$P()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A matrix of P
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-xi"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-xi}{}}}
\subsection{Method \code{xi()}}{
Get exit rates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$xi()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of exit rates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-new}{}}}
\subsection{Method \code{new()}}{
Create an ERHMM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$new(alpha, shape, rate, P, xi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{A vector of initial probability}

\item{\code{shape}}{A vector of shape parameters}

\item{\code{rate}}{A vector of rate parameters}

\item{\code{P}}{A matrix of transition probabilities}

\item{\code{xi}}{An exit rate vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of ERHMM
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-copy"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-copy}{}}}
\subsection{Method \code{copy()}}{
copy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$copy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-size"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-size}{}}}
\subsection{Method \code{size()}}{
The number of components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of components
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-df"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-df}{}}}
\subsection{Method \code{df()}}{
Degrees of freedom
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The degrees of freedom
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-print"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-print}{}}}
\subsection{Method \code{print()}}{
Print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-mmoment"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-mmoment}{}}}
\subsection{Method \code{mmoment()}}{
Marginal moments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$mmoment(k, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{An integer of degree}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of moments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-jmoment"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-jmoment}{}}}
\subsection{Method \code{jmoment()}}{
Joint moments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$jmoment(lag, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lag}}{An integer of lag}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix of moments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-acf"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-acf}{}}}
\subsection{Method \code{acf()}}{
k-lag correlation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$acf(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector for k-lag correlation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-emfit"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-emfit}{}}}
\subsection{Method \code{emfit()}}{
Run EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$emfit(data, options, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{options}}{A list of options}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-init"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-init}{}}}
\subsection{Method \code{init()}}{
Initialize with data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$init(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ERHMMClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ERHMMClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ERHMMClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
