% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gmmpp.R
\name{GMMPPClass}
\alias{GMMPPClass}
\title{GMMPP: Approximation for MAP}
\description{
GMMPP: Approximation for MAP

GMMPP: Approximation for MAP
}
\details{
A point process dominated by a continuous-time Markov chain.
}
\section{Super class}{
\code{\link[mapfit:MAPClass]{mapfit::MAPClass}} -> \code{GMMPPClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GMMPPClass-new}{\code{GMMPPClass$new()}}
\item \href{#method-GMMPPClass-copy}{\code{GMMPPClass$copy()}}
\item \href{#method-GMMPPClass-emfit}{\code{GMMPPClass$emfit()}}
\item \href{#method-GMMPPClass-clone}{\code{GMMPPClass$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="MAPClass" data-id="D0"><a href='../../mapfit/html/MAPClass.html#method-MAPClass-D0'><code>mapfit::MAPClass$D0()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="MAPClass" data-id="D1"><a href='../../mapfit/html/MAPClass.html#method-MAPClass-D1'><code>mapfit::MAPClass$D1()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="MAPClass" data-id="acf"><a href='../../mapfit/html/MAPClass.html#method-MAPClass-acf'><code>mapfit::MAPClass$acf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="MAPClass" data-id="alpha"><a href='../../mapfit/html/MAPClass.html#method-MAPClass-alpha'><code>mapfit::MAPClass$alpha()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="MAPClass" data-id="df"><a href='../../mapfit/html/MAPClass.html#method-MAPClass-df'><code>mapfit::MAPClass$df()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="MAPClass" data-id="init"><a href='../../mapfit/html/MAPClass.html#method-MAPClass-init'><code>mapfit::MAPClass$init()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="MAPClass" data-id="jmoment"><a href='../../mapfit/html/MAPClass.html#method-MAPClass-jmoment'><code>mapfit::MAPClass$jmoment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="MAPClass" data-id="mmoment"><a href='../../mapfit/html/MAPClass.html#method-MAPClass-mmoment'><code>mapfit::MAPClass$mmoment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="MAPClass" data-id="print"><a href='../../mapfit/html/MAPClass.html#method-MAPClass-print'><code>mapfit::MAPClass$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="MAPClass" data-id="size"><a href='../../mapfit/html/MAPClass.html#method-MAPClass-size'><code>mapfit::MAPClass$size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="MAPClass" data-id="xi"><a href='../../mapfit/html/MAPClass.html#method-MAPClass-xi'><code>mapfit::MAPClass$xi()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMMPPClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-GMMPPClass-new}{}}}
\subsection{Method \code{new()}}{
Create a MAP
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMMPPClass$new(alpha, D0, D1, xi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{A vector of initial probability}

\item{\code{D0}}{An infinitesimal generator}

\item{\code{D1}}{An infinitesimal generator}

\item{\code{xi}}{An exit rate vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of MAP
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMMPPClass-copy"></a>}}
\if{latex}{\out{\hypertarget{method-GMMPPClass-copy}{}}}
\subsection{Method \code{copy()}}{
copy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMMPPClass$copy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMMPPClass-emfit"></a>}}
\if{latex}{\out{\hypertarget{method-GMMPPClass-emfit}{}}}
\subsection{Method \code{emfit()}}{
Run EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMMPPClass$emfit(data, options, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{options}}{A list of options}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GMMPPClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GMMPPClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GMMPPClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
