% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_herlang.R
\name{herlang.param}
\alias{herlang.param}
\title{Determine hyper-Erlang parameters}
\usage{
herlang.param(data, shape, ...)
}
\arguments{
\item{data}{A dataframe}

\item{shape}{A vector of shape parameters}

\item{...}{Others}
}
\value{
An instance of HErlang
}
\description{
Determine the hyper-Erlang parameters with k-means.
}
