% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_herlang.R
\name{HErlang}
\alias{HErlang}
\title{Hyper-Erlang distribution}
\description{
Hyper-Erlang distribution

Hyper-Erlang distribution
}
\details{
A mixture of Erlang distributions. A subclass of PH distributions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HErlang-alpha}{\code{HErlang$alpha()}}
\item \href{#method-HErlang-shape}{\code{HErlang$shape()}}
\item \href{#method-HErlang-rate}{\code{HErlang$rate()}}
\item \href{#method-HErlang-new}{\code{HErlang$new()}}
\item \href{#method-HErlang-size}{\code{HErlang$size()}}
\item \href{#method-HErlang-df}{\code{HErlang$df()}}
\item \href{#method-HErlang-moment}{\code{HErlang$moment()}}
\item \href{#method-HErlang-print}{\code{HErlang$print()}}
\item \href{#method-HErlang-pdf}{\code{HErlang$pdf()}}
\item \href{#method-HErlang-cdf}{\code{HErlang$cdf()}}
\item \href{#method-HErlang-ccdf}{\code{HErlang$ccdf()}}
\item \href{#method-HErlang-sample}{\code{HErlang$sample()}}
\item \href{#method-HErlang-emfit}{\code{HErlang$emfit()}}
\item \href{#method-HErlang-init}{\code{HErlang$init()}}
\item \href{#method-HErlang-clone}{\code{HErlang$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-alpha"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-alpha}{}}}
\subsection{Method \code{alpha()}}{
Get alpha
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$alpha()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of alpha
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-shape"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-shape}{}}}
\subsection{Method \code{shape()}}{
Get shape
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$shape()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of shapes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-rate"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-rate}{}}}
\subsection{Method \code{rate()}}{
Get rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$rate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of rates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-new"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-new}{}}}
\subsection{Method \code{new()}}{
Create a hyper-Erlang distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$new(alpha, shape, rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{A vector of initial probability}

\item{\code{shape}}{A vector of shape parameters}

\item{\code{rate}}{A vector of rate parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of HErlang
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-size"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-size}{}}}
\subsection{Method \code{size()}}{
The number of components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of components
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-df"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-df}{}}}
\subsection{Method \code{df()}}{
Degrees of freedom
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The degrees of freedom
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-moment"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-moment}{}}}
\subsection{Method \code{moment()}}{
Moments of HErlang
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$moment(k, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{A value to indicate the degrees of moments. k-th moment}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of moments from 1st to k-th moments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-print"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-print}{}}}
\subsection{Method \code{print()}}{
Print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-pdf}{}}}
\subsection{Method \code{pdf()}}{
PDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$pdf(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of densities.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-cdf}{}}}
\subsection{Method \code{cdf()}}{
CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$cdf(q, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of probabilities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-ccdf"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-ccdf}{}}}
\subsection{Method \code{ccdf()}}{
Complementary CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$ccdf(q, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of probabilies
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-sample"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-sample}{}}}
\subsection{Method \code{sample()}}{
Make a sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$sample(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A sample of HErlang
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-emfit"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-emfit}{}}}
\subsection{Method \code{emfit()}}{
Run EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$emfit(data, options, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{options}}{A list of options}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-init"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-init}{}}}
\subsection{Method \code{init()}}{
Initialize with data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$init(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{...}}{Others}

\item{\code{options}}{A list of options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlang-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HErlang-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlang$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
