% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_herlang.R
\name{HErlangClass}
\alias{HErlangClass}
\title{Hyper-Erlang distribution}
\description{
Hyper-Erlang distribution

Hyper-Erlang distribution
}
\details{
A mixture of Erlang distributions. A subclass of PH distributions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HErlangClass-mixrate}{\code{HErlangClass$mixrate()}}
\item \href{#method-HErlangClass-shape}{\code{HErlangClass$shape()}}
\item \href{#method-HErlangClass-rate}{\code{HErlangClass$rate()}}
\item \href{#method-HErlangClass-new}{\code{HErlangClass$new()}}
\item \href{#method-HErlangClass-copy}{\code{HErlangClass$copy()}}
\item \href{#method-HErlangClass-size}{\code{HErlangClass$size()}}
\item \href{#method-HErlangClass-df}{\code{HErlangClass$df()}}
\item \href{#method-HErlangClass-moment}{\code{HErlangClass$moment()}}
\item \href{#method-HErlangClass-print}{\code{HErlangClass$print()}}
\item \href{#method-HErlangClass-pdf}{\code{HErlangClass$pdf()}}
\item \href{#method-HErlangClass-cdf}{\code{HErlangClass$cdf()}}
\item \href{#method-HErlangClass-ccdf}{\code{HErlangClass$ccdf()}}
\item \href{#method-HErlangClass-sample}{\code{HErlangClass$sample()}}
\item \href{#method-HErlangClass-emfit}{\code{HErlangClass$emfit()}}
\item \href{#method-HErlangClass-init}{\code{HErlangClass$init()}}
\item \href{#method-HErlangClass-clone}{\code{HErlangClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-mixrate"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-mixrate}{}}}
\subsection{Method \code{mixrate()}}{
Get mixrate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$mixrate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of mixrate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-shape"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-shape}{}}}
\subsection{Method \code{shape()}}{
Get shape
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$shape()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of shapes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-rate"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-rate}{}}}
\subsection{Method \code{rate()}}{
Get rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$rate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of rates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-new}{}}}
\subsection{Method \code{new()}}{
Create a hyper-Erlang distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$new(mixrate, shape, rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mixrate}}{A vector of initial probability}

\item{\code{shape}}{A vector of shape parameters}

\item{\code{rate}}{A vector of rate parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of HErlang
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-copy"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-copy}{}}}
\subsection{Method \code{copy()}}{
copy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$copy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-size"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-size}{}}}
\subsection{Method \code{size()}}{
The number of components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of components
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-df"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-df}{}}}
\subsection{Method \code{df()}}{
Degrees of freedom
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The degrees of freedom
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-moment"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-moment}{}}}
\subsection{Method \code{moment()}}{
Moments of HErlang
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$moment(k, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{A value to indicate the degrees of moments. k-th moment}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of moments from 1st to k-th moments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-print"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-print}{}}}
\subsection{Method \code{print()}}{
Print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-pdf}{}}}
\subsection{Method \code{pdf()}}{
PDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$pdf(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of densities.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-cdf}{}}}
\subsection{Method \code{cdf()}}{
CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$cdf(q, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of probabilities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-ccdf"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-ccdf}{}}}
\subsection{Method \code{ccdf()}}{
Complementary CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$ccdf(q, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{A vector of points}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of probabilities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-sample"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-sample}{}}}
\subsection{Method \code{sample()}}{
Make a sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$sample(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A sample of HErlang
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-emfit"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-emfit}{}}}
\subsection{Method \code{emfit()}}{
Run EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$emfit(data, options, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{options}}{A list of options}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-init"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-init}{}}}
\subsection{Method \code{init()}}{
Initialize with data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$init(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{...}}{Others}

\item{\code{options}}{A list of options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HErlangClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HErlangClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HErlangClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
