% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapfit.R
\name{mapfit.point}
\alias{mapfit.point}
\title{MAP fitting with point data}
\usage{
mapfit.point(map, x, intervals, ...)
}
\arguments{
\item{map}{An object of R6 class representing a MAP. The estimation algorithm is selected based on this class.}

\item{x}{A numeric vector of observed point data (event times).}

\item{intervals}{A numeric vector of interval lengths. Optional.}

\item{...}{Additional options for the fitting algorithm.}
}
\value{
A list of class \code{mapfit.result} containing the following components:
\item{model}{An object representing the estimated MAP.}
\item{llf}{The maximized log-likelihood value.}
\item{df}{The degrees of freedom of the fitted model.}
\item{aic}{The Akaike information criterion (AIC) value.}
\item{iter}{The number of iterations performed.}
\item{convergence}{A logical value indicating whether the algorithm converged.}
\item{ctime}{The computation time (user time).}
\item{data}{An object containing the input data set.}
\item{aerror}{The absolute error of the log-likelihood at the final iteration.}
\item{rerror}{The relative error of the log-likelihood at the final iteration.}
\item{options}{A list of options used for the fitting.}
\item{call}{The matched function call.}
}
\description{
Fits a Markovian Arrival Process (MAP) to point data via maximum likelihood estimation.
}
\examples{
## Load trace data
data(BCpAug89)
BCpAug89s <- head(BCpAug89, 50)

## Fit a general MAP
result1 <- mapfit.point(map = map(2), x = cumsum(BCpAug89s))

## Fit a Markov Modulated Poisson Process (MMPP)
result2 <- mapfit.point(map = mmpp(2), x = cumsum(BCpAug89s))

## Fit an Erlang Renewal Hidden Markov Model (ER-HMM)
result3 <- mapfit.point(map = erhmm(3), x = cumsum(BCpAug89s))

## Marginal moments
map.mmoment(k = 3, map = result1$model)
map.mmoment(k = 3, map = result2$model)
map.mmoment(k = 3, map = result3$model)

## Joint moments
map.jmoment(lag = 1, map = result1$model)
map.jmoment(lag = 1, map = result2$model)
map.jmoment(lag = 1, map = result3$model)

## Lag-k autocorrelation
map.acf(map = result1$model)
map.acf(map = result2$model)
map.acf(map = result3$model)

}
