% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_map.R
\name{mmpp}
\alias{mmpp}
\title{Create an MMPP}
\usage{
mmpp(size)
}
\arguments{
\item{size}{An integer for the number of phases}
}
\value{
An instance of MMPP
}
\description{
Create an instance of MMPP (Markov-Modulated Poisson Process)
}
\note{
MMPP is a MAP whose D1 is given by a diagonal matrix.
}
\examples{
## create an MMPP with 5 phases
(param1 <- mmpp(5))

}
