% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ph.R
\name{ph.bidiag}
\alias{ph.bidiag}
\title{Create a bi-diagonal PH distribution}
\usage{
ph.bidiag(size)
}
\arguments{
\item{size}{An integer for the number of phases}
}
\value{
An instance of bi-diagonal PH distribution
}
\description{
Create an instance of bi-diagonal PH distribution.
}
\note{
Bi-diagonal PH distribution is the PH distribution whose infinitesimal
generator is given by a upper bi-diagonal matrix. This is similar to
canonical form 1. But there is no restriction on the order for diagonal
elements.
}
\examples{
## create a bidiagonal PH with 5 phases
(param1 <- ph.bidiag(5))

}
