% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ph.R
\name{ph.mean}
\alias{ph.mean}
\title{Mean of PH distribution}
\usage{
ph.mean(ph, ...)
}
\arguments{
\item{ph}{An instance of PH distribution}

\item{...}{Others}
}
\value{
A value of mean
}
\description{
Compute the mean of a given PH distribution.
}
\examples{
## create PH instance (ex. full PH with 5 phases)
(phdist <- ph(5))

ph.mean(ph)

}
