% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ph.R
\name{ph.tridiag}
\alias{ph.tridiag}
\title{Create a tri-diagonal PH distribution}
\usage{
ph.tridiag(size)
}
\arguments{
\item{size}{An integer for the number of phases}
}
\value{
An instance of tri-diagonal PH distribution
}
\description{
Create an instance of tri-diagonal PH distribution.
}
\note{
Tri-diagonal PH distribution is the PH distribution whose infinitesimal
generator is given by a tri-diagonal matrix (band matrix).
}
\examples{
## create a tridiagonal PH with 5 phases
(param1 <- ph.tridiag(5))

}
