% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ph.R
\name{ph.var}
\alias{ph.var}
\title{Variance of PH distribution}
\usage{
ph.var(ph, ...)
}
\arguments{
\item{ph}{An instance of PH distribution}

\item{...}{Others}
}
\value{
A value of variance
}
\description{
Compute the variance of a given PH distribution.
}
\examples{
## create PH instance (ex. full PH with 5 phases)
(phdist <- ph(5))

ph.var(ph)

}
