% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phfit.R
\name{phfit.group}
\alias{phfit.group}
\title{PH fitting with grouped data}
\usage{
phfit.group(ph, counts, breaks, intervals, instants, ...)
}
\arguments{
\item{ph}{An object of R6 class for PH distributions. The estimation algorithm is selected based on this class.}

\item{counts}{A numeric vector of counts in each interval.}

\item{breaks}{A numeric vector of boundaries for the intervals. Equivalent to \code{c(0, cumsum(intervals))}.
If omitted, defaults to \code{0:length(counts)}.}

\item{intervals}{A numeric vector of interval lengths. Equivalent to \code{diff(breaks)}.
If omitted, defaults to \code{rep(1, length(counts))}.}

\item{instants}{A numeric vector indicating whether an instantaneous sample is drawn at the end of each interval.
If \code{instants[i] = 1}, a sample is drawn at the right boundary of the \code{i}-th interval.
If omitted, defaults to \code{rep(0L, length(counts))} (no instantaneous samples).}

\item{...}{Additional options for the EM algorithm.}
}
\value{
A list of class \code{phfit.result} with the following components:
\item{model}{An object for the estimated PH distribution.}
\item{llf}{The maximized log-likelihood value.}
\item{df}{The degrees of freedom of the fitted model.}
\item{aic}{The Akaike information criterion (AIC) value.}
\item{iter}{The number of iterations performed.}
\item{convergence}{A logical value indicating whether the algorithm converged.}
\item{ctime}{The computation time (user time).}
\item{data}{An object containing the input data.}
\item{aerror}{The absolute error of the log-likelihood at the final iteration.}
\item{rerror}{The relative error of the log-likelihood at the final iteration.}
\item{options}{A list of options used for the fitting.}
\item{call}{The matched function call.}
}
\description{
Fits a phase-type (PH) distribution to grouped data via maximum likelihood estimation.
}
\note{
This method allows handling of truncated data by using \code{NA} and \code{Inf}:

\code{phfit.group(ph = cf1(5), counts = c(countsdata, NA), breaks = c(breakdata, +Inf))} \\
Here, \code{NA} indicates missing count data, and \code{Inf} represents an open-ended last interval \code{[last break point, infinity)}.
}
\examples{
## Generate a sample
wsample <- rweibull(n = 100, shape = 2, scale = 1)
wgroup <- hist(x = wsample, breaks = "fd", plot = FALSE)

## Fit a general PH distribution
result1 <- phfit.group(ph = ph(2), counts = wgroup$counts, breaks = wgroup$breaks)

## Fit a Canonical Form 1 (CF1) distribution
result2 <- phfit.group(ph = cf1(2), counts = wgroup$counts, breaks = wgroup$breaks)

## Fit a hyper-Erlang distribution
result3 <- phfit.group(ph = herlang(3), counts = wgroup$counts, breaks = wgroup$breaks)

## Calculate statistics
ph.mean(result1$model)
ph.var(result2$model)
ph.moment(5, result3$model)

}
