% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phfit.R
\name{phfit.interval}
\alias{phfit.interval}
\title{PH fitting with interval data}
\usage{
phfit.interval(ph, data, weights, ...)
}
\arguments{
\item{ph}{An object of R6 class for PH distributions. The estimation algorithm is selected based on this class.}

\item{data}{A list consisting of point data and interval data.}

\item{weights}{A numeric vector of weights.}

\item{...}{Additional options for the fitting algorithm.}
}
\value{
A list of class \code{phfit.result} with the following components:
\item{model}{An object for the estimated PH distribution.}
\item{llf}{The maximized log-likelihood value.}
\item{df}{The degrees of freedom of the fitted model.}
\item{aic}{The Akaike information criterion (AIC) value.}
\item{iter}{The number of iterations performed.}
\item{convergence}{A logical value indicating whether the algorithm converged.}
\item{ctime}{The computation time (user time).}
\item{data}{An object containing the input data.}
\item{aerror}{The absolute error of the log-likelihood at the final iteration.}
\item{rerror}{The relative error of the log-likelihood at the final iteration.}
\item{options}{A list of options used for the fitting.}
\item{call}{The matched function call.}
}
\description{
Fits a phase-type (PH) distribution to interval data via maximum likelihood estimation.
}
\note{
\code{herlang} cannot be directly used with \code{phfit.interval} yet.
Use general PH (\code{ph}) or Canonical Form 1 (\code{cf1}) models.
}
