% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phfit.R
\name{phfit.point}
\alias{phfit.point}
\title{PH fitting with point data}
\usage{
phfit.point(ph, x, weights, ...)
}
\arguments{
\item{ph}{An object of R6 class for PH distributions. The estimation algorithm is selected based on this class.}

\item{x}{A numeric vector of observed point data (e.g., event times).}

\item{weights}{A numeric vector of weights associated with each point. If omitted, equal weights are assumed.}

\item{...}{Additional options for the fitting algorithm.}
}
\value{
A list of class \code{phfit.result} with the following components:
\item{model}{An object for the estimated PH distribution.}
\item{llf}{The maximized log-likelihood value.}
\item{df}{The degrees of freedom of the fitted model.}
\item{aic}{The Akaike information criterion (AIC) value.}
\item{iter}{The number of iterations performed.}
\item{convergence}{A logical value indicating whether the algorithm converged.}
\item{ctime}{The computation time (user time).}
\item{data}{An object containing the input data.}
\item{aerror}{The absolute error of the log-likelihood at the final iteration.}
\item{rerror}{The relative error of the log-likelihood at the final iteration.}
\item{options}{A list of options used for the fitting.}
\item{call}{The matched function call.}
}
\description{
Fits a phase-type (PH) distribution to point data via maximum likelihood estimation.
}
\examples{
## Generate a sample
wsample <- rweibull(n = 100, shape = 2, scale = 1)

## Fit a general PH distribution
result1 <- phfit.point(ph = ph(2), x = wsample)

## Fit a Canonical Form 1 (CF1) distribution
result2 <- phfit.point(ph = cf1(2), x = wsample)

## Fit a hyper-Erlang distribution
result3 <- phfit.point(ph = herlang(3), x = wsample)

## Calculate statistics
ph.mean(result1$model)
ph.var(result2$model)
ph.moment(5, result3$model)

}
