% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ph.R
\name{pphase}
\alias{pphase}
\title{Distribution function of PH distribution}
\usage{
pphase(q, ph, lower.tail = TRUE, log.p = FALSE, ...)
}
\arguments{
\item{q}{A numeric vector of quantiles.}

\item{ph}{An instance of PH distribution.}

\item{lower.tail}{logical; if TRUE, probabilities are P(X <= x), otherwise P(X > x).}

\item{log.p}{logical; if TRUE, probabilities p are returned as log(p).}

\item{...}{Others}
}
\value{
A vector of densities
}
\description{
Compute the cumulative distribution fuction (c.d.f.) for a given PH distribution
}
\examples{
## create PH instance (ex. full PH with 5 phases)
(phdist <- ph(5))

x <- runif(10) # points of quantiles
pphase(x, ph)

}
