% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ph.R
\name{rphase}
\alias{rphase}
\title{Sampling of PH distributions}
\usage{
rphase(n, ph, ...)
}
\arguments{
\item{n}{An integer of the number of samples.}

\item{ph}{An instance of PH distribution.}

\item{...}{Ohters}
}
\value{
A vector of samples.
}
\description{
Generate a sample from a given PH distribution.
}
\examples{
## create PH instance (ex. full PH with 5 phases)
(phdist <- ph(5))

## Generate 10 samples
rphase(10, ph)

}
