% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cf1.R
\name{cf1}
\alias{cf1}
\title{Create CF1}
\usage{
cf1(size, alpha, rate)
}
\arguments{
\item{size}{An integer of the number of phases}

\item{alpha}{A vector of initial probabilities}

\item{rate}{A vector of rates}
}
\value{
An instance of CF1.
}
\description{
Create an instance of CF1.
}
\examples{
## create a CF1 with 5 phases
(param1 <- cf1(5))

## create a CF1 with 5 phases
(param1 <- cf1(size=5))

## create a CF1 with specific parameters
(param2 <- cf1(alpha=c(1,0,0), rate=c(1.0,2.0,3.0)))

}
