% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cf1.R
\name{CF1Class}
\alias{CF1Class}
\title{Canonical phase-type distribution}
\description{
Canonical phase-type distribution

Canonical phase-type distribution
}
\details{
A continuous distribution dominated by a continuous-time Markov chain.
A random time is given by an absorbing time. In the CF1 (canonical form 1),
the infinitesimal generator is given by a bi-diagonal matrix, and whose order
is determined by the ascending order.
}
\section{Super class}{
\code{\link[mapfit:GPHClass]{mapfit::GPHClass}} -> \code{CF1Class}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CF1Class-rate}{\code{CF1Class$rate()}}
\item \href{#method-CF1Class-new}{\code{CF1Class$new()}}
\item \href{#method-CF1Class-copy}{\code{CF1Class$copy()}}
\item \href{#method-CF1Class-print}{\code{CF1Class$print()}}
\item \href{#method-CF1Class-sample}{\code{CF1Class$sample()}}
\item \href{#method-CF1Class-emfit}{\code{CF1Class$emfit()}}
\item \href{#method-CF1Class-init}{\code{CF1Class$init()}}
\item \href{#method-CF1Class-clone}{\code{CF1Class$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPHClass" data-id="Q"><a href='../../mapfit/html/GPHClass.html#method-GPHClass-Q'><code>mapfit::GPHClass$Q()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPHClass" data-id="alpha"><a href='../../mapfit/html/GPHClass.html#method-GPHClass-alpha'><code>mapfit::GPHClass$alpha()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPHClass" data-id="ccdf"><a href='../../mapfit/html/GPHClass.html#method-GPHClass-ccdf'><code>mapfit::GPHClass$ccdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPHClass" data-id="cdf"><a href='../../mapfit/html/GPHClass.html#method-GPHClass-cdf'><code>mapfit::GPHClass$cdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPHClass" data-id="df"><a href='../../mapfit/html/GPHClass.html#method-GPHClass-df'><code>mapfit::GPHClass$df()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPHClass" data-id="moment"><a href='../../mapfit/html/GPHClass.html#method-GPHClass-moment'><code>mapfit::GPHClass$moment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPHClass" data-id="pdf"><a href='../../mapfit/html/GPHClass.html#method-GPHClass-pdf'><code>mapfit::GPHClass$pdf()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPHClass" data-id="size"><a href='../../mapfit/html/GPHClass.html#method-GPHClass-size'><code>mapfit::GPHClass$size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mapfit" data-topic="GPHClass" data-id="xi"><a href='../../mapfit/html/GPHClass.html#method-GPHClass-xi'><code>mapfit::GPHClass$xi()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF1Class-rate"></a>}}
\if{latex}{\out{\hypertarget{method-CF1Class-rate}{}}}
\subsection{Method \code{rate()}}{
Get rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF1Class$rate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An instance of rate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF1Class-new"></a>}}
\if{latex}{\out{\hypertarget{method-CF1Class-new}{}}}
\subsection{Method \code{new()}}{
Create a CF1
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF1Class$new(alpha, rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{A vector of initial probability}

\item{\code{rate}}{A vector of rates}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of CF1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF1Class-copy"></a>}}
\if{latex}{\out{\hypertarget{method-CF1Class-copy}{}}}
\subsection{Method \code{copy()}}{
copy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF1Class$copy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF1Class-print"></a>}}
\if{latex}{\out{\hypertarget{method-CF1Class-print}{}}}
\subsection{Method \code{print()}}{
Print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF1Class$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF1Class-sample"></a>}}
\if{latex}{\out{\hypertarget{method-CF1Class-sample}{}}}
\subsection{Method \code{sample()}}{
Generate a sample of CF1
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF1Class$sample(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A sample of CF1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF1Class-emfit"></a>}}
\if{latex}{\out{\hypertarget{method-CF1Class-emfit}{}}}
\subsection{Method \code{emfit()}}{
Run EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF1Class$emfit(data, options, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{options}}{A list of options}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF1Class-init"></a>}}
\if{latex}{\out{\hypertarget{method-CF1Class-init}{}}}
\subsection{Method \code{init()}}{
Initialize with data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF1Class$init(data, options, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{options}}{A list of options}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF1Class-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CF1Class-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF1Class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
