% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_phase.R
\name{data.frame.phase.group}
\alias{data.frame.phase.group}
\title{Create group data for phase}
\usage{
data.frame.phase.group(counts, breaks, intervals, instants)
}
\arguments{
\item{counts}{A vector of the number of samples}

\item{breaks}{A vector of break points}

\item{intervals}{A vector of differences of time}

\item{instants}{A vector meaning whether a sample is observed at the end of break.}
}
\value{
A dataframe
}
\description{
Provide the data.frame for group data.
}
\examples{
dat <- data.frame.phase.group(counts=c(1,2,1,1,0,0,1,4))
print(dat)
mean(dat)

}
