% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ph.R
\name{GPH}
\alias{GPH}
\title{General phase-type distribution}
\description{
General phase-type distribution

General phase-type distribution
}
\details{
A continuous distribution dominated by a continuous-time Markov chain.
A random time is given by an absorbing time.
}
\note{
This function provides the values of p.d.f. for PH distribution with
the uniformization technique.

This function provides the values of c.d.f. for PH distribution with
the uniformization technique.

This function provides the values of complementary c.d.f. for
PH distribution with the uniformization technique.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GPH-alpha}{\code{GPH$alpha()}}
\item \href{#method-GPH-Q}{\code{GPH$Q()}}
\item \href{#method-GPH-xi}{\code{GPH$xi()}}
\item \href{#method-GPH-new}{\code{GPH$new()}}
\item \href{#method-GPH-size}{\code{GPH$size()}}
\item \href{#method-GPH-df}{\code{GPH$df()}}
\item \href{#method-GPH-moment}{\code{GPH$moment()}}
\item \href{#method-GPH-print}{\code{GPH$print()}}
\item \href{#method-GPH-pdf}{\code{GPH$pdf()}}
\item \href{#method-GPH-cdf}{\code{GPH$cdf()}}
\item \href{#method-GPH-ccdf}{\code{GPH$ccdf()}}
\item \href{#method-GPH-sample}{\code{GPH$sample()}}
\item \href{#method-GPH-emfit}{\code{GPH$emfit()}}
\item \href{#method-GPH-init}{\code{GPH$init()}}
\item \href{#method-GPH-clone}{\code{GPH$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-alpha"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-alpha}{}}}
\subsection{Method \code{alpha()}}{
Get alpha
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$alpha()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of alpha
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-Q"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-Q}{}}}
\subsection{Method \code{Q()}}{
Get Q
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$Q()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A matrix of Q
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-xi"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-xi}{}}}
\subsection{Method \code{xi()}}{
Get xi
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$xi()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of xi
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-new"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-new}{}}}
\subsection{Method \code{new()}}{
Create a GPH
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$new(alpha, Q, xi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{A vector of initial probability}

\item{\code{Q}}{An infinitesimal generator}

\item{\code{xi}}{An exit rate vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of GPH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-size"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-size}{}}}
\subsection{Method \code{size()}}{
The number of phases
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of phases
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-df"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-df}{}}}
\subsection{Method \code{df()}}{
Degrees of freedom
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The degrees of freedom
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-moment"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-moment}{}}}
\subsection{Method \code{moment()}}{
Moments of GPH
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$moment(k, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{A value to indicate the degrees of moments. k-th moment}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of moments from 1st to k-th moments
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-print"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-print}{}}}
\subsection{Method \code{print()}}{
Print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-pdf}{}}}
\subsection{Method \code{pdf()}}{
PDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$pdf(x, poisson.eps = 1e-08, ufactor = 1.01, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of points}

\item{\code{poisson.eps}}{A value of tolerance error for uniformization}

\item{\code{ufactor}}{A value of uniformization factor}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of densities.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-cdf}{}}}
\subsection{Method \code{cdf()}}{
CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$cdf(x, poisson.eps = 1e-08, ufactor = 1.01, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of points}

\item{\code{poisson.eps}}{A value of tolerance error for uniformization}

\item{\code{ufactor}}{A value of uniformization factor}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of probabilities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-ccdf"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-ccdf}{}}}
\subsection{Method \code{ccdf()}}{
Complementary CDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$ccdf(x, poisson.eps = 1e-08, ufactor = 1.01, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A vector of points}

\item{\code{poisson.eps}}{A value of tolerance error for uniformization}

\item{\code{ufactor}}{A value of uniformization factor}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of probabilies
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-sample"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-sample}{}}}
\subsection{Method \code{sample()}}{
Make a sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$sample(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A sample of GPH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-emfit"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-emfit}{}}}
\subsection{Method \code{emfit()}}{
Run EM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$emfit(data, options, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{options}}{A list of options}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-init"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-init}{}}}
\subsection{Method \code{init()}}{
Initialize with data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$init(data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe}

\item{\code{...}}{Others}

\item{\code{options}}{A list of options}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPH-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GPH-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GPH$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
