% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapfit-package.R
\docType{package}
\name{mapfit-package}
\alias{mapfit}
\alias{mapfit-package}
\title{mapfit: PH/MAP Parameter Estimation}
\description{
Estimation methods for phase-type distribution (PH) and Markovian arrival process (MAP) from empirical data (point and grouped data) and density function. The tool is based on the following researches: Okamura et al. (2009) \doi{10.1109/TNET.2008.2008750}, Okamura and Dohi (2009) \doi{10.1109/QEST.2009.28}, Okamura et al. (2011) \doi{10.1016/j.peva.2011.04.001}, Okamura et al. (2013) \doi{10.1002/asmb.1919}, Horvath and Okamura (2013) \doi{10.1007/978-3-642-40725-3_10}, Okamura and Dohi (2016) \doi{10.15807/jorsj.59.72}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/okamumu/mapfit}
  \item Report bugs at \url{https://github.com/okamumu/mapfit/issues}
}

}
\author{
\strong{Maintainer}: Hiroyuki Okamura \email{okamu@hiroshima-u.ac.jp} (\href{https://orcid.org/0000-0001-6881-0593}{ORCID})

}
\keyword{package}
