% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phfit_mm.R
\name{phfit.3mom}
\alias{phfit.3mom}
\title{PH fitting with three moments}
\usage{
phfit.3mom(
  m1,
  m2,
  m3,
  method = c("Osogami06", "Bobbio05"),
  max.phase = 50,
  epsilon = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{m1}{A value of the first moment.}

\item{m2}{A value of the second moment.}

\item{m3}{A value of the third moment.}

\item{method}{The name of moment matching method.}

\item{max.phase}{An integer for the maximum number of phases in the method
"Osogami06".}

\item{epsilon}{A value of precision in the method "Osogami06".}
}
\value{
An object of GPH.
}
\description{
Estimates PH parameters from three moments.
}
\note{
The method "Osogami06" checks the first three moments on whether there
exists a PH whose three moments match to them. In such case, the method
"Bobbio05" often returns an error.
}
\examples{
## Three moment matching
## Moments of Weibull(shape=2, scale=1); (0.886227, 1.0, 1.32934)
(result1 <- phfit.3mom(0.886227, 1.0, 1.32934))

## Three moment matching
## Moments of Weibull(shape=2, scale=1); (0.886227, 1.0, 1.32934)
(result2 <- phfit.3mom(0.886227, 1.0, 1.32934, method="Bobbio05"))

## mean
ph.mean(result1)
ph.mean(result2)

## variance
ph.var(result1)
ph.var(result2)

## up to 5 moments 
ph.moment(5, result1)
ph.moment(5, result2)

}
\references{
Osogami, T. and Harchol-Balter, M. (2006)
Closed Form Solutions for Mapping General Distributions to Minimal PH Distributions.
\emph{Performance Evaluation}, \bold{63}(6), 524--552.

Bobbio, A., Horvath, A. and Telek, M. (2005)
Matching Three Moments with Minimal Acyclic Phase Type Distributions.
\emph{Stochastic Models}, \bold{21}(2-3), 303--326.
}
