% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phfit.R
\name{phfit.surv}
\alias{phfit.surv}
\title{PH fitting with left-truncated and right-censored data}
\usage{
phfit.surv(ph, x, delta, tau, ...)
}
\arguments{
\item{ph}{An object of R6 class for PH. The estimation algorithm is selected depending on this class.}

\item{x}{A vector for point data (censoring time or event time).}

\item{delta}{A vector of indicators whether x is censoring time or not. If delta=1, the corresponding x is the censoring time
If delta=0, the corresponding x is the event time.}

\item{tau}{A vector of left-truncation time points. If tau is missing, all the left-truncation times are NA (no truncation).}

\item{...}{Further options for fitting methods.}
}
\value{
Returns a list with components, which is an object of S3 class \code{phfit.result};
\item{model}{an object for estimated PH class.}
\item{llf}{a value of the maximum log-likelihood.}
\item{df}{a value of degrees of freedom of the model.}
\item{aic}{a value of Akaike information criterion.}
\item{iter}{the number of iterations.}
\item{convergence}{a logical value for the convergence of estimation algorithm.}
\item{ctime}{computation time (user time).}
\item{data}{an object for data class}
\item{aerror}{a value of absolute error for llf at the last step of algorithm.}
\item{rerror}{a value of relative error for llf at the last step of algorithm.}
\item{options}{a list of optiions used for fitting.}
\item{call}{the matched call.}
}
\description{
Estimates PH parameters from left-truncated and right-censored data.
}
